move doc generator code
