+++ checkov/terraform/graph_builder/graph_components/blocks.py
@@ -1,10 +1,12 @@
 from __future__ import annotations
 
 import os
-from typing import Union, Dict, Any, List, Optional, Set, TYPE_CHECKING
+from typing import Union, Dict, Any, List, Optional, Set, TYPE_CHECKING, cast
 import dpath
 import re
 
+from checkov.common.graph.graph_builder import CustomAttributes
+from checkov.common.graph.graph_builder.utils import calculate_hash
 from checkov.common.runners.base_runner import strtobool
 from checkov.common.typing import TFDefinitionKeyType
 from checkov.terraform.graph_builder.utils import INTERPOLATION_EXPR
@@ -90,6 +92,56 @@ class TerraformBlock(Block):
         return self.name == other.name and self.config == other.config and self.path == other.path and \
             self.block_type == other.block_type and self.attributes == other.attributes and \
             self.id == other.id and self.has_dynamic_block == other.has_dynamic_block and self.source == other.source
+
+    def get_attribute_dict(self, add_hash: bool = True) -> Dict[str, Any]:
+        """
+           :return: map of all the block's native attributes (from the source file),
+           combined with the attributes generated by the module builder.
+           If the attributes are not a primitive type, they are converted to strings.
+           """
+        base_attributes = self.get_base_attributes()
+        self.get_origin_attributes(base_attributes)
+
+        if hasattr(self, "module_dependency") and hasattr(self, "module_dependency_num"):
+            base_attributes[CustomAttributes.MODULE_DEPENDENCY] = self.module_dependency
+            base_attributes[CustomAttributes.MODULE_DEPENDENCY_NUM] = self.module_dependency_num
+
+        if self.changed_attributes:
+            # add changed attributes only for calculating the hash
+            base_attributes["changed_attributes"] = sorted(self.changed_attributes.keys())
+
+        if self.breadcrumbs:
+            sorted_breadcrumbs = dict(sorted(self.breadcrumbs.items()))
+            base_attributes[CustomAttributes.RENDERING_BREADCRUMBS] = sorted_breadcrumbs
+
+        if hasattr(self, 'foreach_attrs'):
+            base_attributes[CustomAttributes.FOREACH_ATTRS] = self.foreach_attrs
+
+        if hasattr(self, 'source_module_object'):
+            base_attributes[CustomAttributes.SOURCE_MODULE_OBJECT] = self.source_module_object
+
+        if add_hash:
+            base_attributes[CustomAttributes.HASH] = calculate_hash(base_attributes)
+
+        if self.block_type == BlockType.DATA:
+            base_attributes[CustomAttributes.RESOURCE_TYPE] = f'data.{self.id.split(".")[0]}'
+
+        if self.block_type == BlockType.MODULE:
+            # since module names are user defined we are just setting 'module' as resource type for easier searching
+            base_attributes[CustomAttributes.RESOURCE_TYPE] = "module"
+
+        if self.block_type == BlockType.PROVIDER:
+            # provider_name is always a string, base_attributes needs better typing pipenv run mypy
+            provider_name = cast(str, base_attributes[CustomAttributes.BLOCK_NAME])
+            provider_type = provider_name.split(".")[0]
+            # ex: provider.aws
+            base_attributes[CustomAttributes.RESOURCE_TYPE] = f"provider.{provider_type}"
+
+        if "changed_attributes" in base_attributes:
+            # removed changed attributes if it was added previously for calculating hash.
+            del base_attributes["changed_attributes"]
+
+        return base_attributes
 
     def add_module_connection(self, attribute_key: str, vertex_id: int) -> None:
         self.module_connections.setdefault(attribute_key, []).append(vertex_id)
